requirements.txt
asgiref==3.7.2
asttokens==2.4.0
backcall==0.2.0
colorama==0.4.6
decorator==5.1.1
Django==3.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
executing==1.2.0
ipython==8.15.0
jedi==0.19.0
matplotlib-inline==0.1.6
mysqlclient==2.2.0
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.0
prompt-toolkit==3.0.39
pure-eval==0.2.2
Pygments==2.16.1
PyJWT==2.8.0
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
traitlets==5.9.0
wcwidth==0.2.6
